Update DropzoneAreaBase.js to allow custom content
Hi @Yuvaleros, thank you for this great plugin!  I am working in a typescript project, and I want to customize the JSX/HTML content of what's inside the dropzone.  I see you can do so with the `dropzoneText` prop, but you have it typed as a string.  I want to put a full-blown JSX element in there.  It actually does work to do so, like this:

```js
export default function App() {
  return (
    <div className="App">
      <DropzoneArea
        dropzoneText={
          <>
            <h2>Something</h2>
            <p>something</p>
            <Button variant="outlined">Custom stuff</Button>
          </>
        }
      />
    </div>
  );
}
```

This renders properly, as expected based on how the `DropzoneAreaBase` component renders the `dropzoneText` prop.  ***However***, in TypeScript projects, I get an error, because this is not a string, but rather a react node.  I can fake it by casing it `as` a string:

```ts
          dropzoneText={
            (
              <>
                <h2>Something</h2>
                <p>something</p>
              </>
            ) as unknown as string
          }
```

But this seems silly.  Can we change this to allow any type of react node, so that typescript is happy?  Thanks again.